# Reduce generics for various Has* traits