--- a/libafl/src/corpus/minimizer.rs
+++ b/libafl/src/corpus/minimizer.rs
@@ -81,29 +81,26 @@ where
 /// corpus that exercise all the requested features (e.g. all the coverage seen so far)
 /// prioritizing [`Testcase`]`s` using [`FavFactor`]
 #[derive(Debug, Clone)]
-pub struct MinimizerCorpusScheduler<C, CS, F, I, M, R, S>
+pub struct MinimizerCorpusScheduler<CS, F, I, M, S>
 where
     CS: CorpusScheduler<I, S>,
     F: FavFactor<I>,
     I: Input,
     M: AsSlice<usize> + SerdeAny + HasRefCnt,
-    S: HasCorpus<C, I> + HasMetadata,
-    C: Corpus<I>,
+    S: HasCorpus<I> + HasMetadata,
 {
     base: CS,
     skip_non_favored_prob: u64,
-    phantom: PhantomData<(C, F, I, M, R, S)>,
+    phantom: PhantomData<(F, I, M, S)>,
 }
 
-impl<C, CS, F, I, M, R, S> CorpusScheduler<I, S> for MinimizerCorpusScheduler<C, CS, F, I, M, R, S>
+impl<CS, F, I, M, S> CorpusScheduler<I, S> for MinimizerCorpusScheduler<CS, F, I, M, S>
 where
     CS: CorpusScheduler<I, S>,
     F: FavFactor<I>,
     I: Input,
     M: AsSlice<usize> + SerdeAny + HasRefCnt,
-    S: HasCorpus<C, I> + HasMetadata + HasRand<R>,
-    C: Corpus<I>,
-    R: Rand,
+    S: HasCorpus<I> + HasMetadata + HasRand,
 {
     /// Add an entry to the corpus and return its index
     fn on_add(&self, state: &mut S, idx: usize) -> Result<(), Error> {
@@ -145,15 +142,13 @@ where
     }
 }
 
-impl<C, CS, F, I, M, R, S> MinimizerCorpusScheduler<C, CS, F, I, M, R, S>
+impl<CS, F, I, M, S> MinimizerCorpusScheduler<CS, F, I, M, S>
 where
     CS: CorpusScheduler<I, S>,
     F: FavFactor<I>,
     I: Input,
     M: AsSlice<usize> + SerdeAny + HasRefCnt,
-    S: HasCorpus<C, I> + HasMetadata + HasRand<R>,
-    C: Corpus<I>,
-    R: Rand,
+    S: HasCorpus<I> + HasMetadata + HasRand,
 {
     /// Update the `Corpus` score using the `MinimizerCorpusScheduler`
     #[allow(clippy::unused_self)]
@@ -284,10 +279,10 @@ where
 }
 
 /// A [`MinimizerCorpusScheduler`] with [`LenTimeMulFavFactor`] to prioritize quick and small [`Testcase`]`s`.
-pub type LenTimeMinimizerCorpusScheduler<C, CS, I, M, R, S> =
-    MinimizerCorpusScheduler<C, CS, LenTimeMulFavFactor<I>, I, M, R, S>;
+pub type LenTimeMinimizerCorpusScheduler<CS, I, M, S> =
+    MinimizerCorpusScheduler<CS, LenTimeMulFavFactor<I>, I, M, S>;
 
 /// A [`MinimizerCorpusScheduler`] with [`LenTimeMulFavFactor`] to prioritize quick and small [`Testcase`]`s`
 /// that exercise all the entries registered in the [`MapIndexesMetadata`].
-pub type IndexesLenTimeMinimizerCorpusScheduler<C, CS, I, R, S> =
-    MinimizerCorpusScheduler<C, CS, LenTimeMulFavFactor<I>, I, MapIndexesMetadata, R, S>;
+pub type IndexesLenTimeMinimizerCorpusScheduler<CS, I, S> =
+    MinimizerCorpusScheduler<CS, LenTimeMulFavFactor<I>, I, MapIndexesMetadata, S>;