--- a/libafl/src/events/simple.rs
+++ b/libafl/src/events/simple.rs
@@ -11,10 +11,7 @@ use crate::{
 };
 use alloc::{string::ToString, vec::Vec};
 #[cfg(feature = "std")]
-use core::{
-    marker::PhantomData,
-    sync::atomic::{compiler_fence, Ordering},
-};
+use core::sync::atomic::{compiler_fence, Ordering};
 #[cfg(feature = "std")]
 use serde::{de::DeserializeOwned, Serialize};
 
@@ -232,11 +229,9 @@ where
 #[cfg(feature = "std")]
 #[allow(clippy::default_trait_access)]
 #[derive(Debug, Clone)]
-pub struct SimpleRestartingEventManager<'a, C, I, MT, S, SC, SP>
+pub struct SimpleRestartingEventManager<I, MT, SP>
 where
-    C: Corpus<I>,
     I: Input,
-    S: Serialize,
     SP: ShMemProvider,
     MT: Monitor, //CE: CustomEvent<I, OT>,
 {
@@ -244,17 +239,12 @@ where
     simple_event_mgr: SimpleEventManager<I, MT>,
     /// [`StateRestorer`] for restarts
     staterestorer: StateRestorer<SP>,
-    /// Phantom data
-    _phantom: PhantomData<&'a (C, I, S, SC)>,
 }
 
 #[cfg(feature = "std")]
-impl<'a, C, I, MT, S, SC, SP> EventFirer<I>
-    for SimpleRestartingEventManager<'a, C, I, MT, S, SC, SP>
+impl<I, MT, SP> EventFirer<I> for SimpleRestartingEventManager<I, MT, SP>
 where
-    C: Corpus<I>,
     I: Input,
-    S: Serialize,
     SP: ShMemProvider,
     MT: Monitor, //CE: CustomEvent<I, OT>,
 {
@@ -264,10 +254,8 @@ where
 }
 
 #[cfg(feature = "std")]
-impl<'a, C, I, MT, S, SC, SP> EventRestarter<S>
-    for SimpleRestartingEventManager<'a, C, I, MT, S, SC, SP>
+impl<I, MT, S, SP> EventRestarter<S> for SimpleRestartingEventManager<I, MT, SP>
 where
-    C: Corpus<I>,
     I: Input,
     S: Serialize,
     SP: ShMemProvider,
@@ -282,10 +270,8 @@ where
 }
 
 #[cfg(feature = "std")]
-impl<'a, C, E, I, MT, S, SC, SP, Z> EventProcessor<E, I, S, Z>
-    for SimpleRestartingEventManager<'a, C, I, MT, S, SC, SP>
+impl<E, I, S, SP, MT, Z> EventProcessor<E, I, S, Z> for SimpleRestartingEventManager<I, MT, SP>
 where
-    C: Corpus<I>,
     I: Input,
     S: Serialize,
     SP: ShMemProvider,
@@ -297,10 +283,8 @@ where
 }
 
 #[cfg(feature = "std")]
-impl<'a, C, E, I, MT, S, SC, SP, Z> EventManager<E, I, S, Z>
-    for SimpleRestartingEventManager<'a, C, I, MT, S, SC, SP>
+impl<E, I, S, SP, MT, Z> EventManager<E, I, S, Z> for SimpleRestartingEventManager<I, MT, SP>
 where
-    C: Corpus<I>,
     I: Input,
     S: Serialize,
     SP: ShMemProvider,
@@ -309,24 +293,18 @@ where
 }
 
 #[cfg(feature = "std")]
-impl<'a, C, I, MT, S, SC, SP> ProgressReporter<I>
-    for SimpleRestartingEventManager<'a, C, I, MT, S, SC, SP>
+impl<I, MT, SP> ProgressReporter<I> for SimpleRestartingEventManager<I, MT, SP>
 where
     I: Input,
-    C: Corpus<I>,
-    S: Serialize,
     SP: ShMemProvider,
     MT: Monitor, //CE: CustomEvent<I, OT>,
 {
 }
 
 #[cfg(feature = "std")]
-impl<'a, C, I, MT, S, SC, SP> HasEventManagerId
-    for SimpleRestartingEventManager<'a, C, I, MT, S, SC, SP>
+impl<I, MT, SP> HasEventManagerId for SimpleRestartingEventManager<I, MT, SP>
 where
-    C: Corpus<I>,
     I: Input,
-    S: Serialize,
     SP: ShMemProvider,
     MT: Monitor,
 {
@@ -337,12 +315,9 @@ where
 
 #[cfg(feature = "std")]
 #[allow(clippy::type_complexity, clippy::too_many_lines)]
-impl<'a, C, I, MT, S, SC, SP> SimpleRestartingEventManager<'a, C, I, MT, S, SC, SP>
+impl<'a, I, MT, SP> SimpleRestartingEventManager<I, MT, SP>
 where
-    C: Corpus<I>,
     I: Input,
-    S: DeserializeOwned + Serialize + HasCorpus<C, I> + HasSolutions<SC, I>,
-    SC: Corpus<I>,
     SP: ShMemProvider,
     MT: Monitor, //TODO CE: CustomEvent,
 {
@@ -351,7 +326,6 @@ where
         Self {
             staterestorer,
             simple_event_mgr: SimpleEventManager::new(monitor),
-            _phantom: PhantomData {},
         }
     }
 
@@ -359,7 +333,10 @@ where
     /// This [`EventManager`] is simple and single threaded,
     /// but can still used shared maps to recover from crashes and timeouts.
     #[allow(clippy::similar_names)]
-    pub fn launch(mut monitor: MT, shmem_provider: &mut SP) -> Result<(Option<S>, Self), Error> {
+    pub fn launch<S>(mut monitor: MT, shmem_provider: &mut SP) -> Result<(Option<S>, Self), Error>
+    where
+        S: DeserializeOwned + Serialize + HasCorpus<I> + HasSolutions<I>,
+    {
         // We start ourself as child process to actually fuzz
         let mut staterestorer = if std::env::var(_ENV_FUZZER_SENDER).is_err() {
             // First, create a place to store state in, for restarts.